+++ internal/repository/project.repository.go
@@ -18,7 +18,7 @@ func (r *ProjectRepository) Create(project *models.Project) error {
 	return r.db.Create(project).Error
 }
 
-func (r *ProjectRepository) GetByID(id uint) (*models.Project, error) {
+func (r *ProjectRepository) GetByID(id uint64) (*models.Project, error) {
 	var project models.Project
 	if err := r.db.First(&project, "id = ?", id).Error; err != nil {
 		return nil, err
@@ -26,7 +26,7 @@ func (r *ProjectRepository) GetByID(id uint) (*models.Project, error) {
 	return &project, nil
 }
 
-func (r *ProjectRepository) CountUserProjects(userID uint) (uint8, error) {
+func (r *ProjectRepository) CountUserProjects(userID uint64) (uint8, error) {
 	var count int64
 	if err := r.db.Model(&models.Project{}).Where("owner_id = ?", userID).Count(&count).Error; err != nil {
 		return 0, err
@@ -34,7 +34,7 @@ func (r *ProjectRepository) CountUserProjects(userID uint) (uint8, error) {
 	return uint8(count), nil
 }
 
-func (r *ProjectRepository) GetByUserID(userID uint) ([]*models.Project, error) {
+func (r *ProjectRepository) GetByUserID(userID uint64) ([]*models.Project, error) {
 	var projects []*models.Project
 	if err := r.db.Where("owner_id = ?", userID).Find(&projects).Error; err != nil {
 		return nil, err
